+++ apps/course-builder-web/src/app/tips/[slug]/edit/page.tsx
@@ -1,12 +1,9 @@
 import * as React from 'react'
 import { notFound } from 'next/navigation'
-import { reprocessTranscript } from '@/app/tips/[slug]/edit/actions'
 import { getAbility } from '@/lib/ability'
 import { getTip } from '@/lib/tips'
 import { getVideoResource } from '@/lib/video-resource'
 import { getServerAuthSession } from '@/server/auth'
-
-import { Button } from '@coursebuilder/ui'
 
 import { EditTipForm } from '../../_components/edit-tip-form'
 
@@ -21,12 +18,7 @@ export default async function TipEditPage({ params }: { params: { slug: string }
     notFound()
   }
 
-  const videoResourceLoader = getVideoResource(
-    tip.videoResourceId,
-    tip.videoResourceId ? [tip.videoResourceId, 'tips', tip._id] : ['tips', tip._id],
-  )
-
-  const reprocessForVideoResourceId = reprocessTranscript.bind(null, { videoResourceId: tip.videoResourceId })
+  const videoResourceLoader = getVideoResource(tip.videoResourceId)
 
   return <EditTipForm key={tip.slug} tip={tip} videoResourceLoader={videoResourceLoader} />
 }
